requirements.txt
fastapi
uvicorn
pypdf
langchain
langchain-huggingface
chromadb                
google-generativeai
sentence-transformers